move preDecoratedEmbed to scripts.js file
